+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 * add tag flag for the run command instead of @+tagname argument
 * add tag flag for the list command instead of @+tagname argument
 * add status for history (success/fail) and collect history only for the run command
+* migrate from pkg_resources to importlib.resources (python min version 3.7)
 
 ## [0.6.0] - 2024-08-19